Add constraints for protocol java generation (#3119)
* Add constraint validation for Protocol Java generation

* Add tests for Enum constraints

* Always Box primitives

* Fix boolean type

* Fix tests

* Add default to Protocol

* Fix validation

* Fix impacted tests

* Fix one validation edge case

* Flip test for constraints

+++ legend-engine-xts-graphQL/legend-engine-xt-graphQL-grammar/src/main/java/org/finos/legend/engine/language/graphQL/grammar/from/GraphQLGrammarParser.java
@@ -473,7 +473,7 @@ public class GraphQLGrammarParser
         if (valueContext.intValue() != null)
         {
             IntValue value = new IntValue();
-            value.value = Integer.parseInt(valueContext.intValue().getText());
+            value.value = Long.parseLong(valueContext.intValue().getText());
             return value;
         }
         if (valueContext.stringValue() != null)
